system: Remember AP watchdog reset
On MT8183, when EC detects a watchdog reset, EC needs to reboot itself
in preparation for the next boot. This means that AP loses the reset
cause (as AP system reset is toggled), and, therefore, we need to save
the reset reason in the EC.

This has been done in CL:1295890 for chip stm32. For the new MT8183
model 'cozmo', we need to do the same for chip it83xx.

BUG=b:184225435
TEST=firmware_EventLog passed on Cozmo
TEST=FW_NAME=cozmo emerge-jacuzzi chromeos-ec
BRANCH=none

Change-Id: I08d165b5dbd8e1d92fad66ef3873fc406ab6a0e0
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2814079
Reviewed-by: Eric Yilun Lin <yllin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2839077
Tested-by: Eric Yilun Lin <yllin@chromium.org>
Commit-Queue: Eric Yilun Lin <yllin@chromium.org>