cleanup: delete unused legacy kconfig implementation
This code is currently unused.  It was from an effort to bring Kconfig
to CrOS EC OS.

Our EC will receive Kconfig by means of Zephyr-based devices, making
the effort to bring it to the legacy OS likely obsolete.  However, if
for some reason we want to bring it to the legacy OS in the future, we
would probably want to use kconfiglib instead of this C
implementation, as Depthcharge and Zephyr have migrated to it, and
Coreboot is in the process of migrating to kconfiglib.

BUG=none
BRANCH=none
TEST=make buildall

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: If53d0c6bd752a0adba775b8638a4f612ccb399ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2824033
Reviewed-by: caveh jalali <caveh@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2839197
Tested-by: Eric Yilun Lin <yllin@chromium.org>
Reviewed-by: Eric Yilun Lin <yllin@chromium.org>
Commit-Queue: Eric Yilun Lin <yllin@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>