Homestar:LED:Use display SoC to control charge LED
Currently, Homestar uses the state of charge provided from the battery.
This isn't the same as the SoC shown on the screen because Powerd
compensates it based on the full factor and the low battery shutdown
threshold.

This change makes Homestar use the display SoC for the charge LED module
so that the charge LED and the display SoC synchronously work.

BUG=b:187539586
TEST=make -j BOARD=homestar
    Verify build on EVT board
BRANCH=Trogdo

Signed-off-by: tongjian <tongjian@huaqin.corp-partner.google.com>
Change-Id: I953aca48d18b32655d82fbd9a8c315534a8326bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2966517
Reviewed-by: Wai-Hong Tam <waihong@google.com>