cr50: switch SPI CS to GPIO mode
To be able to send and receive multiple SPI buffer quantities in a
single SPI transaction it is necessary to control the SPI CS signal
directly, as opposed to connecting it to the SPI controller.

Direct mode allows to keep CS asserted as long as necessary to
transfer the full data blob, size of which might exceed the the size
of the SPI controller buffer.

BUG=b:79492818
TEST=flashrom access still works fine:

  $ flashrom  -p raiden_debug_spi:target=ap --flash-name
  flashrom f10dff7b-dirty on Linux 5.4.0-71-generic (x86_64)
  flashrom is free software, get the source code at https://flashrom.org

  Using clock_gettime for delay loops (clk_id: 1, resolution: 1ns).
  Raiden target: 2
  Found GigaDevice flash chip "GD25Q127C/GD25Q128C" (16384 kB, SPI) on...
  vendor="GigaDevice" name="GD25Q127C/GD25Q128C"

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: Idecf019d3fd19675d7f78e4dc1140106a2112c6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2973580
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
(cherry picked from commit c6095192742c9f59f72a6178ece5a2dd8cea4e43)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3004085
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>