chgstv2/sustainer: Add tests for full battery and AC
Currently, unit tests do not check whether the battery sustainer can be
enabled when the battery is already full or when the AC is already
present.

This patch adds tests which check the battery sustainer can be enabled
when the battery is already full or when the AC is already present.

BUG=b:188457962
BRANCH=None
TEST=run-sbs_charging_v2

Change-Id: I4ec9785554d126baca0b60e889c4a2dabbfb628a
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3869646
Reviewed-by: Jett Rink <jettrink@chromium.org>