drive/charger/isl9421: Fix race condition with discharge_on_ac
There is a race condition with `learn_mode`. This variable is read and
then discharge_on_ac is disabled if it is not set. However it's possible
for isl9241_discharge_on_ac to get called after the branch is taken but
before discharge_on_ac is disabled. This can cause factory USB charging
tests to fail as discharge_on_ac is disabled improperly. This commit
protects the read and branch on `learn_mode` with a mutex to prevent the
above from happening.

BUG=b:214341758
BRANCH=guybrush
TEST=Ran USB left charge test 2000 times

Change-Id: I64cd63fb2761e2d01b3b384d580ab3597870425f
Signed-off-by: Robert Zieba <robertzieba@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3579624
Reviewed-by: Rob Barnes <robbarnes@google.com>
Reviewed-by: Diana Z <dzigterman@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3708625
Tested-by: Zick Wei <zick.wei@quanta.corp-partner.google.com>
Commit-Queue: Zick Wei <zick.wei@quanta.corp-partner.google.com>