trogdor: Pass through AUX termination when DP-Config is acked
The hardware block diagram is like:

          AUX termination
                 |
            +----+--- TCPC MUX --- PPC SW --- Type-C connector (port-0)
DP PHY -- DP MUX
            +----+--- TCPC MUX --- PPC SW --- Type-C connector (port-1)
                 |
          AUX termination

The current implementation muxes the path end-to-end on a HPD high
event.

The dependencies are:
 * Chromebook muxing the AUX channel depends on a HPD high event,
 * DP sink raising HPD depends on a cable connection detected (auto-
   probe),
 * Successful probe of a cable connection depends on AUX termination,
which cause a dead-lock.

This change breaks the dead-lock. During the DP-Config phase, when the
DP-Config VDM is acked, connect the PPC SW and TCPC MUX, but leave the
DP MUX disconnected. So the DP sink can detect the correct AUX
termination. The DP port selection mux is not enabled yet so no signal
passes through to the DP PHY.

As the TCPC MUX is enabled in the DP-Config phase, we can't use the TCPC
MUX status as the condition to check if the DP path is muxable. Change
it to the DP selection mux.

BRANCH=trogdor
BUG=b:247005613
TEST=Together with the next CL.
Connect the device through a Type-C dongle to a monitor and verify
the screen. Try the combinations:
 * Plug C0 monitor -> plug C0 dongle -> screen on C0 monitor
 * Plug C0 dongle -> plug C0 monitor -> screen on C0 monitor
   -> unplug C0 monitor -> plug C0 monitor -> screen on C0 monitor
 * Plug C0 dongle -> plug C0 monitor -> screen on C0 monitor
   -> plug C1 dongle -> -> plug C1 monitor -> no screen on C1 monitor
   -> unplug C0 monitor -> replug C1 monitor -> screen on C0 monitor
 * Plug C0 dongle -> plug C1 dongle -> plug C1 monitor
   -> screen on C1 monitor -> unplug C1 dongle -> plug C0 monitor
   -> screen on C0 monitor

Change-Id: Id24ce4627a1c015ad2dc9748578fa1a6d68122c6
Signed-off-by: Wai-Hong Tam <waihong@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4076560
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
(cherry picked from commit 5a99b9b38428215cce0f3c660f003919652595e2)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4163457