accel_list2dw12: Set high perf mode for ODR >= 50z
Until recently we're used it only as lid accel, where no significant
precision is required. However in Winterhold we want to use it for
body detection. In order to reduce the noise and make it usable for that
purpose we need to set it into "high performance" power mode.
ODR is used to recognize the sensor purpose.
For lid accel it's set to 12.5Hz, assume that if it's at least 50Hz
the high performance is wanted.

BUG=b:262680246
BRANCH=none
TEST=Set ODR to 50Hz, verify high performance mode is on

Change-Id: I8c4decc93d2f2cfcc074cf2399f5cb582eb72696
Signed-off-by: Kornel Dulęba <korneld@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4379149
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
(cherry picked from commit 8b50c762dc2df4d14e7aff4838e2490419be7913)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4465588
Tested-by: Elthan Huang <elthan_huang@compal.corp-partner.google.com>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Elthan Huang <elthan_huang@compal.corp-partner.google.com>