panic: Set watchdog panic reason if old watchdog panic
The watchdog panic info may be set in watchdog handler or during system
pre init after the watchdog reset. It depends on the chip
implementation. For cortex-m on Zephyr, the watchdog panic reason is
not set in the watchdog handler.

The current logic in system pre init only sets the watchdog panic reason
when the current reason is not a watchdog. This means only the first
watchdog is reported. Successive watchdogs are ignored because the panic
flags are never reset. The panic flags will only be reset if a
non-watchdog panic occurs or EC memory cleared (FW update or 0 battery).

This change checks if the current watchdog panic is old. This is
inferred by checking for the PANIC_DATA_FLAG_OLD_HOSTCMD flag, which is
set when the host reads the panic info.

BUG=b:273516527
BRANCH=Zephyer NPCX boards
TEST=Succesive watchdogs are correctly reported in elog
TEST=drivers.system unit test

Change-Id: I91ec1bf75a85ea7fed978895a262ff74e6ac16c7
Signed-off-by: Rob Barnes <robbarnes@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4415329
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5077677