brescue.sh: try closing process using the console TTY device
Quite often by the moment brescue.sh is invoked the console tty device
is connected to a terminal window.

Instead of bailing out in this case try to determine the terminal PID
and kill the process.

In case of success - proceed with the rescue, if the device is still
taken after five seconds - report the problem and exit.

BUG=none

TEST=tried invoking brescue.sh while the terminal session was active,
     observed the following message on the console:

/dev/ttyUSB2 is in use, will try to kill process 1762753 ............ succeeded

     and then a successful rescue.

Change-Id: I51511deb8e4c6352b893bd760fe146863f029ea4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4609646
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>