re-sign-gsc: script for creating node locked RW GSC image copies
This script is supposed to be ran in GSC root directories (for Cr50 or
Ti50), and it expects the GSC binary image to be already built from
the current tree state.

The two command line parameters the script expects are device ID
values of the chip the image is being signed for.

The script determines which GSC type it is signing, finds the
pre-built binary blob, then finds the RW sections in the blob, each
section is extracted, re-signed with the new device IDs and pasted
back into the binary.

Then the updated binary is saved in the local directory. The name of
the saved binary consists of the GSC type, followed by the git tree
state (hash plus optional '-dirty' if the tree has modified files),
followed by the device IDs used during signing.

BUG=None
TEST=successfully built and signed images for Cr50 and Ti50, for both
     slots A and B

     verified proper FIPS initialization on Cr50:
    ...
    [0.130035 FIPS power-up tests completed in 63385]
    [0.131896 FIPS module digest 596701c64af6a646...]
    [0.132785 Running in FIPS 140-2 approved mode]
    ...

Change-Id: I29da1e17c7354ea20d300c88df1c7b2dc0c64568
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4782054
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Vadim Sukhomlinov <sukhomlinov@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>