gsctool: error when using -D option with cr50 image
The GSC type is automatically selected from the image type that it
passed in and it previously overwrote the -D option. Now if a Cr50 image
is passed in with a -D (for dauntless) option, then gsctool will error
instead of trying to flash the cr50 image on a cr50 chip.

BUG=none
TEST=cr50 image with -D now gives an error

Change-Id: Ifbcdc5891924802be156e357eb7cdfd4dac6f039
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4891656
Tested-by: Jett Rink <jettrink@chromium.org>
Commit-Queue: Jett Rink <jettrink@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>