gsctool: consolidate H1 and DT processing.
Using image_header as an indication of the chip gsctool expects to
operate on is a bit misleading, because many commands do not involve
processing a GSC header. On top of that many commands can work with
either GSC chip, in this case setting 'image_header' to Cr50 and
making it a default does not make sense.

Let's have a separate variable tracking the expected GSC device,
making it an enum, which would allow to distinguish three separate
cases: any, H1, DT.

Add a field to the command line options table to mark options which
operate only on certain GSC types. Make sure that in case image
processing is involved, image type matches the device.

On commands which can operate on either chip, when connected over USB
first try H1 and then DT.

BUG=none
TEST=verify that cr50 image update will only scan for H1 chip
     verify that ti50 image update will only scan for DT chip
     verify that gsctool --reboot will first scan for H1, then DT
     verify that gstool -D --reboot will only scan for DT chip

Change-Id: Ie765070765991ab6d6613a4ff732574bfd915fbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4904803
Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Commit-Queue: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>