dojo: increase chipset task stack size
The crashmon reports DOJO has high crash rates in chipset task (id 2).
Increase the stack size to reduce the issue. Note that we don't know
how much extra stack size is needed, so we add supposedly large enough
value.

The crashmon is reporting that DOJO has a high number of crashes in
chipset task (id 2). To reduce the number of crashes, we increase the
stack size. The exact amount of stack size needed is unknown, so we
should add a value that is large enough.

BUG=b:308362805
TEST=build pass

Change-Id: I5848496f9d70dbd533440b3e4afe7b53ec480df7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4990073
Commit-Queue: Eric Yilun Lin <yllin@google.com>
Tested-by: Eric Yilun Lin <yllin@google.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>