cr50: wipe secrets with randoms, don't touch EPS
EPS seed is same for different users, it persists in NV and there is
no reason to wipe it.

Destroy over secrets by overwriting with randoms if possible, and don't
touch length. This may help to avoid corner cases when call to
nvmem_wipe_cache() somehow clashes with TPM2_Clear or something else
which would dump these NV spaces to NVmem, so make it closer to what
TPM2_Clear() does.

These spaces are only recreated on manufacturing in call to
HierarchyPreInstall_Init().

BUG=b:262324344
TEST=make; check that image works

Change-Id: I9f32d776a9362befbbecf9ade3f88e4f28535ee7
Signed-off-by: Vadim Sukhomlinov <sukhomlinov@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5021018
Reviewed-by: Mary Ruthven <mruthven@chromium.org>