cr50: fix in nvmem_erase_tpm_data_selective()
nvmem_erase_tpm_data_selective runs a loop filling a page till nvmem
access tracker moves to next page. However, if on page 7 or higher,
it didn't account for compaction called by setvar().

Refactor how filling page is done so it doesn't depend on setvar() and
amount of variable space left and is much faster as object is created
as deleted in a single shot.

BUG=b:314732719, b:293234946
TEST=make buildall; tpmtest.py -p
make run-nvmem V=1

Change-Id: I00d65234afa88e523e1ff5a12938e6d598722d15
Signed-off-by: Vadim Sukhomlinov <sukhomlinov@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5120797
Reviewed-by: Vadim Bendebury <vbendeb@google.com>