chip: sign node locked images using a Cloud KMS key
This patch relies on the recently added codesigner functionality which
allows the user to sign GSC images with keys hosted in Cloud KMS.

The private key counterpart of cr50_RW-prod.pem.pub was imported into
Cloud KMS and is used for this mode of signing.

build.mk was cleaned up to make the logic easier to follow. No need to
plug the device IDs into the manifest, it's been a while since the
codesigner supports receiving them through the command line.

Note that to reach Cloud KMS the codesigner needs to be provided a
pointer to the connector library, PKCS11_MODULE_PATH is set to point
to the current location, which has not yet been fully settled.

Both local gnubby based and Cloud KMS based signing modes are still
supported, as before they kick in when H1_DEVIDS variable is set. To
enable KMS based mode one needs to declare USE_CLOUD_KMS_SIGNER to be
nonempty before make invocation, the appropriate codesinger command
line options and environment variable settings are added by build.mk
in this case.

BUG=b:329162115
TEST=all the following commands succeded:
  BOARD=cr50 make
  H1_DEVIDS='x y' BOARD=cr50 make # requires signing fob
  USE_CLOUD_KMS_SIGNER=1 H1_DEVIDS='x y' BOARD=cr50 make

  images build with H2_DEVIDS set can be run on the designated H1.

Cq-Depend: chrome-internal:7310978
Change-Id: Id2253a7aab3412264e32f3a59ff343e61a7f0998
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5560450
Tested-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Vadim Sukhomlinov <sukhomlinov@google.com>
Commit-Queue: Vadim Bendebury <vbendeb@google.com>