dsp_comms: Add client/service
Introduce the client and service logic of the DSP comms. The client is
running on the DSP and sends requests over to the service (EC) via an
I2C bus. When a response is ready, the service (EC) will fire an
interrupt to notify the client (DSP). The client (DSP) will then read a
fixed size status message with the response size or a bit signaling a
processing error. If no error, the client (DSP) will initiate a read
for the response.

Currently the following are implemented:
- client (DSP) notify service (EC) of notebook mode change
- client (DSP) requesting CBI values from service (EC)
- service (EC) notifying client (DSP) of lid-open and tablet mode

BUG=b:345530593
TEST=Unit tests and Brox

Change-Id: Id38895202cbdf80553fb73faaba30cb02a5e7a6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5824175
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>