awasuki: modify als count abnormal
Add a bit in the EEPROM to record the normal triggering
of the ALS process. After the ALS is triggered, repeatedly
plug and unplug the AC. When the battery is activated, and
this bit is set. even the ALS process is triggered again,
it will not be counted.

BUG=b:353445852
TEST=Remove the cover, wait for a while, plug in AC,
the count does not increase by two.

Change-Id: I2d06b72009da05a1107bc154f0bf38ca125df021
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5982411
Tested-by: Mike Lee <mike5@huaqin.corp-partner.google.com>
Reviewed-by: Shou-Chieh Hsu <shouchieh@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Ganxiang Wang <wangganxiang@huaqin.corp-partner.google.com>
Commit-Queue: Shou-Chieh Hsu <shouchieh@chromium.org>
(cherry picked from commit 0e70922fb6cc9980ddea400ff149c1c2c10747ea)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6022725