pdc: tps6699x: Mask status_change bit
This CL removes status_change from the interrupt mask. This bit is not
being checked in the callback handler function and it increases the
chances that an subsequent interrupt might be missed by the EC if the
PDC sets another bit in the time between when the EC reads and
subsequently writes the status bits to the interrupt clear register.

BUG=b:345783692,b:373454623
TEST=Validated to trulo with various port partners.

Change-Id: Ia15d00408848279d7c5bc0326432a70c9506bf02
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6001109
Reviewed-by: Ivan Chen <yulunchen@google.com>
Tested-by: Scott Collyer <scollyer@chromium.org>
Commit-Queue: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>