board/hyperdebug: Route all SPI requests through board code
This patch does not create functional change.  The SPI1 and SPI2 are
declared as "use board-specific driver", on top of QSPI.

Add another layer to the board specific driver, to dispatch between the
QSPI driver, which is specific to HyperDebug, and the ordinary driver in
`chip/stm32`.

BUG=b:376508426
TEST=make tast TAST_EXPR=gscdevboard.GSCTPM.spi_get_time  (in ti50/common)

Change-Id: I9411162de634acbcda9306d32214fd4ce5fd2950
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6022030
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Jes Klinke <jbk@chromium.org>
Commit-Queue: Jes Klinke <jbk@chromium.org>
Reviewed-by: Edward Hill <ecgh@chromium.org>