Telith: modify als count abnormal
Add a bit in the EEPROM to record the normal triggering
of the ALS process. After the ALS is triggered, repeatedly
plug and unplug the AC. When the battery is activated, and
this bit is set. even the ALS process is triggered again,
it will not be counted.

BUG=b:377443087
TEST=Remove the cover, wait for a while, plug in AC,
the count does not increase by two.
LOW_COVERAGE_REASON=Tests will be added in b:377443087

Change-Id: Ie544f84269f61edcbec3a8a0fb45f9e72659da08
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6056990
Tested-by: Mike Lee <mike5@huaqin.corp-partner.google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Ganxiang Wang <wangganxiang@huaqin.corp-partner.google.com>
Reviewed-by: Ricky Chang <rickytlchang@chromium.org>
Commit-Queue: Edward Dai <edwarddai@chromium.org>
Reviewed-by: Edward Dai <edwarddai@chromium.org>