test: thermistor: do not set the mock adc voltage reference
All Zephyr ADC APIs are going to be const for the time being, ADC being
changed in http://crrev.com/c/6060062. Since the internal vref is
implemented as a value in the api structure rather than a function call,
this means that it can't be mocked anymore, not without an api change
anyway. Fix our tests by not setting it at all, the rest of the ADC APIs
seems to be handling the conversion correctly anyway.

BUG=none
TEST=./twister -v -s drivers/drivers.default.mock_power.ec_host_cmd

Cq-Depend: chromium:6060062
Change-Id: Iad050d2ca3d1b5270eff8044dd801258d81b83a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6059978
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>