board/hyperdebug: Fast clock output on CN10.31
A single pin has support for outputting the core clock of HyperDebug,
that is, up to 110MHz signal.  We might as well support that, if the
existing pwm command is used to request such a high clock frequency for
that particular pin.

BUG=b:384755415
TEST=Observed 96MHz waveform from HyperDebug on oscilloscope

Change-Id: I411cb85b43fef0aa2fd6453438fd1a237eb4258f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6113973
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Jes Klinke <jbk@chromium.org>
Commit-Queue: Jes Klinke <jbk@chromium.org>
Reviewed-by: Jett Rink <jettrink@google.com>