dsp_comms: Allow remote lid switch on service
When building the service side (EC) we need to forward the lid-switch
signal to the client. This is controlled by the
PLATFORM_EC_DSP_REMOTE_LID_SWITCH kconfig. To enable this on the EC
(which has PLATFORM_EC_LID_SWITCH) we need to modify the dependency.

BUG=b:345530593
TEST=Flash Orisa and test tablet/clamshell mode

Change-Id: I1c640131774f6663e14cb5a5b0967a25bdf528e1

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6134396
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>